2_implementation/backend/learning-service/requirements.txt
# FastAPI 和相關依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 資料庫相關
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# HTTP 客戶端
httpx==0.25.2

# 認證相關
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 日誌和配置
python-dotenv==1.0.0
structlog==23.2.0

# 工具庫
python-dateutil==2.8.2
pytz==2023.3

# 開發和測試工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 健康檢查
healthcheck==1.3.0 